Synchronized with Zedd's latest plugins

+++ HaloOnlineTagTool/TagStructures/ChudWidgetParallaxData.cs
@@ -12,8 +12,8 @@ namespace HaloOnlineTagTool.TagStructures
 	[TagStructure(Class = "cprl", Size = 0x10)]
 	public class ChudWidgetParallaxData
 	{
-		public uint Unknown;
-		public uint Unknown2;
+		public float Unknown;
+		public float Unknown2;
 		public uint Unknown3;
 		public uint Unknown4;
 	}
