+++ HaloOnlineTagTool/TagStructures/InputGlobals.cs
@@ -12,10 +12,10 @@ namespace HaloOnlineTagTool.TagStructures
 	[TagStructure(Class = "inpg", Size = 0x34)]
 	public class InputGlobals
 	{
-		public uint Unknown;
-		public uint Unknown2;
+		public int Unknown;
+		public float Unknown2;
 		public byte[] Unknown3;
 		public byte[] Unknown4;
-		public uint Unknown5;
+		public int Unknown5;
 	}
 }
